Support loading elements from source
Testing HTML templates outside of the container is very useful. Testing
logic that builds and asserting that they render the data you expect is
computationally expensive when navigating through a browser. Having the
ability to test a html fragment in isolaition allows a lower level of
testing to be done efficiently and without introducing unnecessary
wieght/verbosity to higher level tests.