Return `NotFound` for missing/invisible elements
When an element is not found it's nice to be able to express that in a
test. e.g. `expect(page.element).to be_visible`

Prior to this accesing an element that was invisible or not present on
the page at all would have risen a `ElementMissingException`. This now
happens if any method besides those for checking visibility are called.