Fix broken support for dynamic selectors
introducing `PageMagic::Config` had broken the support for dynamic
selectors. There evaluation used to be 'lazy' and in bringing validation
upfront made it possible for selectors to be declared dynamically when
inline element definitions are evaluated at runtime. This commit puts
the validation back to being at runtime and restores the ability to
declare selectors dynamically.