Provide access to all `Capybara::Session` methods
The PageMagic API doesn't completely obscure the capybara API and
exposing methods available on the seession has some advantages. I.e.
Exposing all of the available methods on `Capybara::Session`
means that the `PageMagic::Session` object can be dropped in a direct
replacement in to test suites that have access to this object through
the methods in the `Capybara::DSL`