Remove dependency on Jeweler
The Jeweler dependency was stoping a number of gems being updated as it
was holding certain shared sub dependencies to old version. It wasn't
doing that much and it appears as though this project has gone end of
life. Therefore this project no longer uses this dependency to manage
building the page_magic gem. Instead we use the `gem` command directly
and use the gemspec to drive the Gemfile instead of the otherway round.